+++ intelmq/tests/bots/parsers/cznic/test_parser_haas.py
@@ -40,7 +40,7 @@ INPUT = """[
 
 OUTPUT_0 = {
     '__type': 'Event',
-    'classification.type': 'unauthorized-command',
+    'classification.type': 'system-compromise',
     'protocol.transport': 'tcp',
     'protocol.application': 'ssh',
     'destination.port': 22,
@@ -69,7 +69,7 @@ OUTPUT_1 = {
 
 OUTPUT_2 = {
     '__type': 'Event',
-    'classification.type': 'unauthorized-login',
+    'classification.type': 'system-compromise',
     'protocol.transport': 'tcp',
     'protocol.application': 'ssh',
     'destination.port': 22,