+++ intelmq/tests/bots/parsers/dyn/test_parser.py
@@ -41,7 +41,7 @@ EXAMPLE_EVENTS = [{"feed.url": "http://security-research.dyndns.org/pub/botnet/p
                    "raw": "LyAyNTMwOC5leGFtcGxlLmNvbSBodHRwOi8vMjUzMDguZXhhbXBsZS5jb20vdXJsIG1hbGljaW91cy5leGFtcGxlLm5ldA==",
                    "time.observation": "2015-11-02T13:11:44+00:00",
                    "time.source": "2016-03-15T07:47:49+00:00",
-                   "classification.type": "compromised",
+                   "classification.type": "system-compromise",
                    "source.fqdn": "malicious.example.net",
                    "event_description.text": "host has been compromised and has malicious code infecting users",
                    "__type": "Event"
@@ -66,7 +66,7 @@ EXAMPLE_EVENTS = [{"feed.url": "http://security-research.dyndns.org/pub/botnet/p
                    "raw": "LyAzNjAxNS5leGFtcGxlLmNvbSBodHRwOi8vMzYwMTUuZXhhbXBsZS5jb20vdXJsIG1hbGljaW91czIuZXhhbXBsZS5uZXQ=",
                    "time.observation": "2015-11-02T13:11:44+00:00",
                    "time.source": "2016-03-15T07:47:49+00:00",
-                   "classification.type": "compromised",
+                   "classification.type": "system-compromise",
                    "source.fqdn": "malicious2.example.net",
                    "event_description.text": "host has been compromised and has malicious code infecting users",
                    "__type": "Event"